+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 1.38
 
 - Automated migration for Capacitor 4 Plugins
+- Detect if gem used to install Cocoapods and use it for upgrades
 
 ### Version 1.37